+++ applications/services/desktop/views/desktop_events.h
@@ -34,9 +34,7 @@ typedef enum {
 
     DesktopPinTimeoutExit,
 
-    DesktopDebugEventDeed,
-    DesktopDebugEventWrongDeed,
-    DesktopDebugEventSaveState,
+    DesktopDebugEventToggleDebugMode,
     DesktopDebugEventExit,
 
     DesktopLockMenuEventLock,