+++ CHANGELOG.md
@@ -14,6 +14,7 @@
 * OFW: Fix Felica standard loading from nfc file
 * Apps: **Check out more Apps updates and fixes by following** [this link](https://github.com/xMasterX/all-the-plugins/commits/dev)
 ## Other changes
+* Desktop: Quick Debug mode switch (hold down on desktop, then hold ok on version screen)
 * SubGHz: OFEX support for SecPlus v1 and v2, various fixes (by @Dmitry422 & xMasterX)
 * SubGHz Remote: Add default remote and clear slot features (by @jknlsn)
 * Fix typo in README warning about scammers (PR #931 | by @koterba)